Allow multi-selection per category
When we are inside a context of multi-selection, it is now possible to
select all locations inside a category; it is also possible to deselect
all locations inside a category.
When selecting several locations, the status of the category CheckBox
now reflects how many locations have been selected inside the given
category.
This is a first attempt to fix issue #2